ShoppingProgram/ebay_token_proxy/requirements.txt
Flask==2.3.5
requests==2.32.0
python-dotenv==1.0.0